requirements.txt
streamlit
langchain
langchain_community
langchain_core
langchain_ollama
pdfplumber
faiss-cpu
PyPDF2
chromadb
openai
gradio 
pandas 
json5 
ollama
docx2txt
docx
sentence-transformers
einops
pymupdf
numpy
python-docx
sentence-transformers

